refactor: Community Page 코드 리팩토링
refactor: #67
