+++ src/pages/Community/CommunityEdit.jsx
@@ -1,16 +1,11 @@
 import React, { useState } from 'react';
-import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
-
-import Alert from '../../components/Alert';
+import EditFooter from '../../components/Comunity/EditFooter';
+import EditContent from '../../components/Comunity/EditContent';
 
 const CommunityEdit = () => {
-  const navigate = useNavigate();
   const [fileName, setFileName] = useState('');
   const [previews, setPreviews] = useState([]);
-
-  const content =
-    '해당 페이지를 벗어나면 작성 중인 글이 모두 삭제됩니다 떠나시겠습니까?';
 
   const handleFileChange = (event) => {
     const selectedFiles = Array.from(event.target.files);
@@ -23,61 +18,9 @@ const CommunityEdit = () => {
     <Container>
       <ContentContainer>
         <TopHeader>커뮤니티 글 작성</TopHeader>
-        <Content>
-          <TitleArea>
-            <ContentTitle>제목</ContentTitle>
-            <TitleInput placeholder="제목을 입력하세요." />
-          </TitleArea>
-          <ContentArea>
-            <ContentMain>내용</ContentMain>
-            <ContentInput placeholder="내용을 입력하세요." />
-          </ContentArea>
-          {fileName.length > 0 && (
-            <ImageContainer>
-              {previews.map((src, index) => (
-                <Image key={index} src={src} alt={`preview-${index}`} />
-              ))}
-            </ImageContainer>
-          )}
-        </Content>
-        <ButtonContainer>
-          <ImageInput>
-            사진
-            <input
-              type="file"
-              multiple
-              onChange={handleFileChange}
-              accept="image/*"
-            />
-          </ImageInput>
-          <RightButtons>
-            <CancelButton onClick={() => navigate('/community')}>
-              취소
-            </CancelButton>
-            <SubmitButton>작성</SubmitButton>
-          </RightButtons>
-        </ButtonContainer>
-        <CatuionContainer>
-          <CautionText>
-            * 과도한 비방 및 욕설이 포함된 게시글은 신고에 의해 무통보 삭제될 수
-            있습니다.
-          </CautionText>
-          <CautionText>
-            * 초상권, 저작권 침해 및 기타 위법한 게시글은 관리자에 의해 무통보
-            삭제될 수 있습니다.
-          </CautionText>
-        </CatuionContainer>
+        <EditContent fileName={fileName} previews={previews} />
+        <EditFooter handleFileChange={handleFileChange} />
       </ContentContainer>
-      {/* <Alert
-        message={content}
-        showButtons={true}
-        RefuseText={'떠나기'}
-        OkayText={'남기'}
-        ContainerWidth={280}
-        ContainerHeight={108}
-        AlertWidth={392}
-        AlertHeight={280}
-      /> */}
     </Container>
   );
 };
@@ -102,216 +45,6 @@ const ContentContainer = styled.div`
   margin-left: 235px;
   margin-right: 205px;
   height: 100%;
-`;
-
-const Content = styled.div`
-  margin-top: 44px;
-  padding-left: 50px;
-  padding-right: 50px;
-  border: 2px solid #9819c3;
-  border-radius: 10px;
-
-  width: 900px;
-  height: 100%;
-`;
-
-const ContentTitle = styled.div`
-  color: #fff;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-`;
-
-const TitleArea = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  border-bottom: 1.5px solid #b2b2b2;
-
-  padding-top: 21px;
-  padding-bottom: 14px;
-  padding-left: 9px;
-`;
-
-const TitleInput = styled.input`
-  border: transparent;
-  background-color: transparent;
-  outline: none;
-
-  color: white;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-
-  &::placeholder {
-    color: #4d4d4d;
-    font-size: 14px;
-  }
-
-  width: 92%;
-  height: 100%;
-`;
-
-const ContentArea = styled.div`
-  display: flex;
-  justify-content: space-between;
-
-  padding-top: 15px;
-  margin-bottom: 49px;
-  padding-left: 9px;
-
-  display: flex;
-`;
-
-const ContentInput = styled.textarea`
-  background-color: transparent;
-  border: transparent;
-  resize: none;
-  outline: none;
-
-  color: white;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-
-  &::placeholder {
-    color: #4d4d4d;
-    font-size: 14px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-  }
-
-  width: 92%;
-  height: 390px;
-`;
-
-const ContentMain = styled.div`
-  color: #fff;
-  font-size: 18px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: normal;
-`;
-
-const ButtonContainer = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  margin-top: 21px;
-  margin-left: 40px;
-  margin-right: 14px;
-`;
-
-const ImageInput = styled.label`
-  display: inline-block;
-  background-color: transparent;
-  border: 2px solid #9819c3;
-  color: #ffffff;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-  text-align: center;
-
-  &:hover {
-    background-color: #9819c3;
-    color: #fff;
-  }
-
-  input[type='file'] {
-    display: none;
-  }
-`;
-
-const RightButtons = styled.div`
-  display: flex;
-  gap: 10px;
-`;
-
-const CancelButton = styled.button`
-  background-color: transparent;
-  border: 2px solid #9819c3;
-  color: #ffffff;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-
-  &:hover {
-    background-color: #9819c3;
-    color: #fff;
-  }
-`;
-
-const SubmitButton = styled.button`
-  background-color: #9819c3;
-  border: none;
-  color: #fff;
-  padding: 10px 20px;
-  border-radius: 5px;
-  cursor: pointer;
-
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 600;
-  line-height: normal;
-
-  &:hover {
-    background-color: #7c16a6;
-  }
-`;
-
-const CatuionContainer = styled.div`
-  margin-left: 40px;
-  margin-top: 20px;
-
-  color: white;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: normal;
-`;
-
-const CautionText = styled.div`
-  margin-bottom: 5px;
-`;
-
-const ImageContainer = styled.div`
-  display: flex;
-  flex-direction: row;
-  gap: 20px;
-  margin-left: 20px;
-  margin-bottom: 46px;
-
-  width: 900px;
-  height: 200px;
-`;
-
-const Image = styled.img`
-  background-color: white;
-  border-radius: 7px;
-  width: 200px;
-  height: 200px;
-`;
-
-const PreviewImage = styled.img`
-  width: 100px;
-  height: 100px;
-  object-fit: cover;
-  margin: 10px;
-  border-radius: 5px;
-  border: 2px solid #9819c3;
 `;
 
 export default CommunityEdit;
